+++ src/main/java/org/sosly/arcaneadditions/spells/components/TransfuseComponent.java
@@ -25,6 +25,7 @@ import com.mna.factions.Factions;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.entity.LivingEntity;
+import org.sosly.arcaneadditions.entities.ai.config.FamiliarAIConfig;
 
 public class TransfuseComponent extends SpellEffect implements IDamageComponent {
     public TransfuseComponent(ResourceLocation guiIcon) {
@@ -39,7 +40,7 @@ public class TransfuseComponent extends SpellEffect implements IDamageComponent 
         LivingEntity livingSource = null;
         float damage = mods.getValue(Attribute.DAMAGE) * GeneralConfig.getDamageMultiplier();
         float magnitude = mods.getValue(Attribute.MAGNITUDE);
-        float healing = damage * (magnitude / 5);
+        float healing = damage * (magnitude / FamiliarAIConfig.TRANSFUSE_EFFICIENCY_DIVISOR);
 
         if (target.isLivingEntity() && target.getLivingEntity() != null) {
             livingTarget = target.getLivingEntity();