+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Familiar AI behavior settings to server config for movement, spellcasting, and wandering
+- Familiar healing rate and interval settings to server config
 
 ### Fixed
 - Added a loot table to the scribe's bench so that it properly drops when broken